update:編集機能を実装

+++ src/App.js
@@ -1,20 +1,23 @@
 import React, { Component } from 'react'
 
 const App = () => {
-  return <Counter />
+  return <ToDo />
 }
 
-class Counter extends Component{
+class ToDo extends Component{
   constructor(props) {
     super(props);
     this.state = {
       value: 0,
       text: '',
       isLoaded: false,
-      items: []
+      items: [],
+      updateId: '',
+      updateTitle: ''
     };
     this.handleChange = this.handleChange.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
+    this.updateTitleChange = this.updateTitleChange.bind(this);
   }
   //componentDidMount()はコンポーネントがマウントされた（ツリーに挿入された）直後に呼び出される
   componentDidMount() { //render直後に行いたい処理を書くところ
@@ -48,8 +51,44 @@ class Counter extends Component{
       .then(console.log)
       .catch(console.error);
   }
+  setUpdateItem(id,title) {
+    this.setState({
+      updateId: id,
+      updateTitle: title
+    })
+  }
+  resetUpdateItem() {
+    this.setState({
+      updateId: '',
+      updateTitle: ''
+    })
+  }
   updateItem(id) {
-    alert(`id:${id}のアイテムの編集に成功しました！`);
+    const method = "PATCH"
+    const headers = {
+      'Accept': 'application/json',
+      'Content-Type': 'application/json'
+    };
+    const body = JSON.stringify({
+      "title": this.state.updateTitle
+    });
+    fetch(`http://localhost:3000/api/v1/posts/${id}`, { method, headers, body })
+      .then(
+        (res) => {
+          console.log(res.stats)
+          this.initItems()
+          alert(`id:${id}のアイテムの編集に成功しました！`);
+          this.setState({
+            updateId: ''
+          })
+          return res.json()
+        })
+      .then(
+        console.log
+      )
+      .catch((err) => {
+        console.log(`err:${err}`);
+      });
   }
 
   //Stateの値を更新する場合、this.setSate()で更新する
@@ -68,6 +107,9 @@ class Counter extends Component{
 
   handleChange(event) {
     this.setState({ text: event.target.value });
+  }
+  updateTitleChange(event) {
+    this.setState({ updateTitle: event.target.value });
   }
 
   handleSubmit(event) {
@@ -106,15 +148,25 @@ class Counter extends Component{
         </div>
         <form onSubmit={this.handleSubmit}>
           <label>
-            <input type="text" value={this.state.text} onChange={this.handleChange} />
+            <input type="text" placeholder="新規タスクを入力"  value={this.state.text} onChange={this.handleChange} />
           </label>
           <input type="submit" value="送信" />
         </form>
         {this.state.items.map((item) => {
-          return <div key={item.id}>{item.title}
-            <button onClick={() => { this.deleteItem(item.id) }}>削除</button>
-            <button onClick={() => { this.updateItem(item.id) }}>編集</button>
-          </div>
+          if (this.state.updateId === item.id) {
+            return <div key={item.id}>
+              <input type="text" placeholder="編集後のタスクを入力" value={this.state.updateTitle} onChange={this.updateTitleChange}/>
+              <button onClick={() => { this.updateItem(item.id) }}>OK</button>
+              <button onClick={() => { this.resetUpdateItem() }}>キャンセル</button>
+            </div>
+          }
+          else {
+            return <div key={item.id}>
+              {item.title}
+              <button onClick={() => { this.deleteItem(item.id) }}>削除</button>
+              <button onClick={() => this.setUpdateItem(item.id,item.title)}>編集</button>
+            </div>
+          }
         })}
       </div>
     );
